fix: footer and some little other thing
