+++ apps/nuxt/src/app.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { MetaSeo } from '@/tools/metaSeo'
 import { useNavigationStore } from './stores/navigation'
 import TeeFooter from './components/TeeFooter.vue'
 import Translation from '@/tools/translation'
@@ -16,21 +17,21 @@ onBeforeMount(() => {
 // cf: https://stackoverflow.com/questions/69495211/vue3-route-query-empty
 navigationStore.setRouter(router)
 navigationStore.setRoute(route)
+
+useSeoMeta(MetaSeo.default())
 </script>
 
 <template>
   <div>
+    <TeeHeader />
+    <NuxtPage />
+    <div class="fr-mt-0v">
+      <TeeFooter />
+    </div>
     <ClientOnly>
       <TeeDsfrConsent />
       <TeeDsfrPersonalizeConsent />
       <TeeRegisterModal v-if="navigationStore.hasRegisterModal" />
     </ClientOnly>
-    <TeeHeader />
-
-    <NuxtPage />
-
-    <div class="fr-mt-0v">
-      <TeeFooter />
-    </div>
   </div>
 </template>
