+++ apps/nuxt/src/assets/scss/footer.scss
@@ -1,8 +1,6 @@
 @use 'tool';
 
 .fr-footer {
-  box-shadow: none !important;
-  
   &__content {
     flex-basis: fit-content;
     margin-left: 0;
