+++ apps/nuxt/src/assets/scss/grid.scss
@@ -57,10 +57,6 @@
   z-index: 1000;
 }
 
-.fr-position--relative { 
+.fr-position--relative {
   position: relative;
 }
-
-.fr-position--absolute {
-  position: absolute;
-}
