+++ apps/nuxt/src/components/project/list/ProjectCard.vue
@@ -52,11 +52,9 @@ const priorityTag: string = 'A FAIRE EN PRIORITÉ'
 
 const navigationStore = useNavigationStore()
 const navigation = new Navigation()
-const isCatalog = navigation.isCatalogProjects()
-const isHomePage = navigation.isByRouteName(RouteName.Homepage)
 const getRouteToProjectDetail = (project: ProjectType): RouteLocationRaw => {
   return {
-    name: isCatalog || isHomePage ? RouteName.CatalogProjectDetail : RouteName.ProjectResultDetail,
+    name: navigation.isCatalogProjects() || navigation.isHomepage() ? RouteName.CatalogProjectDetail : RouteName.ProjectResultDetail,
     params: { projectSlug: project.slug },
     query: navigationStore.query
   }
