+++ apps/nuxt/src/tools/metaSeo.ts
@@ -12,7 +12,7 @@ export class MetaSeo {
   }
 
   static readonly default = () => {
-    return this._buildMeta(this._title(this._defaultTitle), this._image(), this._defaultDescription)
+    return this._buildMeta(this._title(), this._image(), this._defaultDescription)
   }
 
   private static readonly _title = (title?: string) => {
@@ -24,8 +24,10 @@ export class MetaSeo {
   }
 
   private static readonly _image = (image?: string) => {
-    image = image?.charAt(0) === '/' ? image : `/${image}`
-    return image ?? this._defaultImage
+    if (!image) {
+      return this._defaultImage
+    }
+    return image?.charAt(0) === '/' ? image : `/${image}`
   }
 
   private static _buildMeta(title: string, image: string, description?: string): UseSeoMetaInput {
